requirements.txt
substrate_interface==0.13.9
pinatapy_vourhey==0.1.3
scalecodec==0.11.9
ipfshttpclient==0.7.0
PyYAML==5.4.1
substrateinterface==0.9.3
